requirements.txt
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1